requirements.txt
flask
pandas
joblib
xgboost
lightgbm
scikit-learn
python-dotenv
psycopg2
